issue-3405: Spike C4 — Cross-reference map for counter/integrity global callsites (#3522)

+++ textquest/src/travel/graph.rs
@@ -663,7 +663,7 @@ mod tests {
             zone_points.push(point(
                 &format!("z{index}"),
                 "LiveEra",
-                index,
+                index as i32,
                 0,
                 0,
             ));
